refactor: rename "authentication.py" to "security.py"
The features from this file can be reused for encryption,
so keeping the "authentication" name would be misleading.

Ref: NCSDK-28258
