+++ suit_generator/suit/manifest.py
@@ -28,7 +28,7 @@ from suit_generator.suit.types.common import (
     SuitBchar,
     cbstr,
 )
-from suit_generator.suit.authentication import SuitDigest
+from suit_generator.suit.security import SuitDigest
 from suit_generator.suit.types.keys import (
     suit_parameter_vendor_identifier,
     suit_parameter_class_identifier,
